Add JRuby extension wrappers for all metric types
When marking a meter thousands of times per second looking up an object proxy for the meter takes measurable amounts of time, and methods are called by reflection. By wrapping the metric objects with something that JRuby understands this lookup can be minimized.

The #to_h implementations are not time critical and much easier to implement in Ruby, so they have been left as-is.