Make it possible to create gauges that return Java types
Some reporters, e.g. metrics-graphite, inspects the return type of gauges to see if it's one of a list of expected types, because that's the only way they can ensure that they can correctly format the value. Since IRubyObject very rarely works there needs to be a way to create types gauges.