Use Jetty for HTTP and expose separate Rack app
This replaces the old Rack based HTTP server with a Jetty-based one. A
Rack app is still exposed, but the handling of it is left to the
user. This way, the Rack app can be mounted anywhere in an existing
application.

Moreover, and more importantly, is that both of these now use the
Metrics JSON serializations library, which means that it is possible to
get very consistent reporting even with applications using vanilla Metrics.