Replace #to_h with #to_json
I assume that #to_h is only used as an intermediate step in the
serialization for most users, why I think it may be ok to skip it. By
using Metrics JSON library, it's easier to maintain a consistent format
across applications. Or maybe rather; it is difficult not to.